+++ phantom4/dosfunc.h
@@ -3,8 +3,8 @@
 
 #include "dosdata.h"
 
-extern ulong dos_ftime(void);
+extern uint32_t dos_ftime(void);
 extern void set_sft_owner(SFTREC_PTR sft);
-extern int is_a_character_device(uint dos_ds);
+extern int is_a_character_device(uint16_t dos_ds);
 
 #endif /* _DOSFUNC_H */
